split source
