+++ scoreWDLstat.cpp
@@ -1,3 +1,5 @@
+#include "scoreWDLstat.hpp"
+
 #include <atomic>
 #include <chrono>
 #include <cmath>
@@ -20,39 +22,6 @@ using json   = nlohmann::json;
 
 using namespace chess;
 
-enum class Result { WIN = 'W', DRAW = 'D', LOSS = 'L' };
-
-struct ResultKey {
-    Result white;
-    Result black;
-};
-
-struct Key {
-    Result outcome;             // game outcome from PoV of side to move
-    int move, material, score;  // move number, material count, engine's eval
-    bool operator==(const Key &k) const {
-        return outcome == k.outcome && move == k.move && material == k.material && score == k.score;
-    }
-    operator std::size_t() const {
-        // golden ratio hashing, thus 0x9e3779b9
-        std::uint32_t hash = static_cast<int>(outcome);
-        hash ^= move + 0x9e3779b9 + (hash << 6) + (hash >> 2);
-        hash ^= material + 0x9e3779b9 + (hash << 6) + (hash >> 2);
-        hash ^= score + 0x9e3779b9 + (hash << 6) + (hash >> 2);
-        return hash;
-    }
-    operator std::string() const {
-        return "('" + std::string(1, static_cast<char>(outcome)) + "', " + std::to_string(move) +
-               ", " + std::to_string(material) + ", " + std::to_string(score) + ")";
-    }
-};
-
-// overload the std::hash function for Key
-template <>
-struct std::hash<Key> {
-    std::size_t operator()(const Key &k) const { return static_cast<std::size_t>(k); }
-};
-
 // unordered map to count (outcome, move, material, score) tuples in pgns
 using map_t = std::unordered_map<Key, int>;
 
@@ -62,50 +31,6 @@ using map_meta = std::unordered_map<std::string, json>;
 std::atomic<std::size_t> total_chunks = 0;
 
 namespace analysis {
-
-/// @brief Custom stof implementation to avoid locale issues, once clang supports std::from_chars
-/// for floats this can be removed
-/// @param str
-/// @return
-float fast_stof(const char *str) {
-    float result   = 0.0f;
-    int sign       = 1;
-    int decimal    = 0;
-    float fraction = 1.0f;
-
-    // Handle sign
-    if (*str == '-') {
-        sign = -1;
-        str++;
-    } else if (*str == '+') {
-        str++;
-    }
-
-    // Convert integer part
-    while (*str >= '0' && *str <= '9') {
-        result = result * 10.0f + (*str - '0');
-        str++;
-    }
-
-    // Convert decimal part
-    if (*str == '.') {
-        str++;
-        while (*str >= '0' && *str <= '9') {
-            result = result * 10.0f + (*str - '0');
-            fraction *= 10.0f;
-            str++;
-        }
-        decimal = 1;
-    }
-
-    // Apply sign and adjust for decimal
-    result *= sign;
-    if (decimal) {
-        result /= fraction;
-    }
-
-    return result;
-}
 
 /// @brief Magic value for fishtest pgns, ~1.2 million keys
 static constexpr int map_size = 1200000;
@@ -286,27 +211,6 @@ void ana_files(map_t &map, const std::vector<std::string> &files, const std::str
 
 }  // namespace analysis
 
-/// @brief Get all files from a directory.
-/// @param path
-/// @param recursive
-/// @return
-[[nodiscard]] std::vector<std::string> get_files(const std::string &path, bool recursive = false) {
-    std::vector<std::string> files;
-
-    for (const auto &entry : fs::directory_iterator(path)) {
-        if (fs::is_regular_file(entry)) {
-            if (entry.path().extension() == ".pgn") {
-                files.push_back(entry.path().string());
-            }
-        } else if (recursive && fs::is_directory(entry)) {
-            auto subdir_files = get_files(entry.path().string(), true);
-            files.insert(files.end(), subdir_files.begin(), subdir_files.end());
-        }
-    }
-
-    return files;
-}
-
 [[nodiscard]] map_meta get_metadata(const std::vector<std::string> &file_list,
                                     bool allow_duplicates) {
     map_meta meta_map;
@@ -386,29 +290,6 @@ void filter_files_sprt(std::vector<std::string> &file_list, const map_meta &meta
     file_list.erase(std::remove_if(file_list.begin(), file_list.end(), pred), file_list.end());
 }
 
-/// @brief Split into successive n-sized chunks from pgns.
-/// @param pgns
-/// @param target_chunks
-/// @return
-[[nodiscard]] std::vector<std::vector<std::string>> split_chunks(
-    const std::vector<std::string> &pgns, int target_chunks) {
-    const int chunks_size = (pgns.size() + target_chunks - 1) / target_chunks;
-
-    auto begin = pgns.begin();
-    auto end   = pgns.end();
-
-    std::vector<std::vector<std::string>> chunks;
-
-    while (begin != end) {
-        auto next =
-            std::next(begin, std::min(chunks_size, static_cast<int>(std::distance(begin, end))));
-        chunks.push_back(std::vector<std::string>(begin, next));
-        begin = next;
-    }
-
-    return chunks;
-}
-
 void process(const std::vector<std::string> &files_pgn, map_t &pos_map,
              const std::string &regex_engine, const map_meta &meta_map, bool fix_fens) {
     // Create more chunks than threads to prevent threads from idling.
@@ -476,14 +357,6 @@ void save(const map_t &pos_map, const std::string &json_filename) {
 
     std::cout << "Wrote " << total << " scored positions to " << json_filename << " for analysis."
               << std::endl;
-}
-
-bool find_argument(const std::vector<std::string> &args,
-                   std::vector<std::string>::const_iterator &pos, std::string_view arg,
-                   bool without_parameter = false) {
-    pos = std::find(args.begin(), args.end(), arg);
-
-    return pos != args.end() && (without_parameter || std::next(pos) != args.end());
 }
 
 void print_usage(char const *program_name) {
